remove unused permissions
