+++ juntagrico/fixtures/test/members.json
@@ -90,11 +90,6 @@
                 "specialroles"
             ],
             [
-                "is_operations_group",
-                "juntagrico",
-                "specialroles"
-            ],
-            [
                 "depot_list_notification",
                 "juntagrico",
                 "specialroles"
