+++ juntagrico_legacy/views.py
@@ -1,6 +1,7 @@
 from io import BytesIO
 
 from django.contrib.auth.decorators import permission_required, login_required
+from django.core.exceptions import ValidationError
 from django.http import HttpResponse
 from django.shortcuts import render, get_object_or_404
 from django.utils.translation import gettext as _
@@ -16,7 +17,8 @@ from juntagrico.util import addons, temporal, return_to_previous_location
 from juntagrico.util.management import create_subscription_parts
 
 from juntagrico.util.management_list import get_changedate
-from juntagrico.view_decorators import primary_member_of_subscription
+from juntagrico.view_decorators import primary_member_of_subscription, using_change_date
+from juntagrico.views_subscription import error_page
 
 from juntagrico_legacy.util.xls import generate_excel
 from juntagrico_legacy.dao.subscriptiondao import SubscriptionDao
@@ -56,6 +58,28 @@ def part_canceledlist(request):
     render_dict = get_changedate(request)
     changedlist = SubscriptionPartDao.canceled_parts_for_active_subscriptions()
     return subscription_management_list(changedlist, render_dict, 'management_lists/part_canceledlist.html', request)
+
+
+@permission_required('juntagrico.change_subscription')
+@using_change_date
+def activate_subscription(request, change_date, subscription_id):
+    subscription = get_object_or_404(Subscription, id=subscription_id)
+    try:
+        subscription.activate(change_date)
+    except ValidationError as e:
+        return error_page(request, e.message)
+    return return_to_previous_location(request)
+
+
+@permission_required('juntagrico.change_subscription')
+@using_change_date
+def deactivate_subscription(request, change_date, subscription_id):
+    subscription = get_object_or_404(Subscription, id=subscription_id)
+    try:
+        subscription.deactivate(change_date)
+    except ValidationError as e:
+        return error_page(request, e.message)
+    return return_to_previous_location(request)
 
 
 @permission_required('juntagrico.change_subscriptionpart')
